requirements.txt
antlr4-python3-runtime==4.8
backcall==0.2.0
certifi==2020.6.20
colorama==0.4.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
future==0.18.2
hydra-core==1.0.3
importlib-resources==3.2.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.3.0
matplotlib==3.3.2
numpy==1.19.2
omegaconf==2.0.3
pandas==1.1.3
parso==0.7.1
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
PyYAML==5.3.1
pyzmq==19.0.2
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
six==1.15.0
sklearn==0.0
statsmodels==0.12.0
threadpoolctl==2.1.0
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
tornado==6.0.4
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5
